fix(planner): cutover review pass 5 — Bug 1 (blocker) + Issue 2 + minor + nit
Pass 5 verdict: "Approve with fixes for Bug #1 before merge; Issues #2-4
acceptable as follow-ups."

Bug 1 (blocker)

- appStateSchema.scheduledSubSchema.timeSeconds: was z.number().min(0),
  now z.number().int().min(0). validateScheduledSubs (validation.ts)
  rejects non-integer timeSeconds and throws on the next save. A backup
  containing fractional timeSeconds would pass schema validation in the
  restore path but produce a permanently un-saveable game. The live
  timer also compares newTime >= s.timeSeconds against whole-second
  ticks, so a fractional sub might never fire. Schema integer
  constraint matches the DataStore reality.

Issue 2 (graduated to in-PR — small enough to land cleanly)

- PlanningModal.handleEditorApplied: stamps PlanningSession.appliedAt
  on every successful Apply via fire-and-forget mutateAsync. The field
  was defined, mapped to applied_at in the DB, threaded through every
  rename/duplicate/save call — but never actually written. Logger.warn
  on failure rather than blocking the undo banner; Apply has already
  succeeded against the games (the snapshot proves it).

Issue 3 (documented, not restructured)

- SyncedDataStore.setActiveSession: added comment describing the
  staleness window between the unguarded `before` read and the locked
  setActiveSession call. Single-user-single-tab usage makes this race
  effectively unreachable; restructuring LocalDataStore.setActiveSession
  to return the diff instead would touch the API signature and is out
  of scope for the cutover.

Issue 4 (already documented)

- savePlanningSession SELECT round-trip: existing inline TODO already
  flags the COALESCE-on-conflict fix. No change.

Minor 5

- validatePlanningSession: isActive type check. Every other required
  field is validated; isActive wasn't. A corrupt cloud row with
  isActive: "yes" (string) would have surfaced in the session list as
  a non-boolean. One-liner closes the surface; tiny blast radius given
  the DB NOT NULL DEFAULT false constraint, but matches the validator's
  intent.

Nit 8

- planMinutesAggregate.ts: comment explaining why startingXISize is
  sampled once outside the gameIds loop (the editor replicates one
  PlanDraft across every game; per-game drafts would require moving
  the sample inside).

Skipped

- Minor 6 (captureApplyableFields shallow-clone) — Claude confirmed
  the existing comment is load-bearing; no code change needed.
- Minor 7 (PLANNING_SESSIONS_KEY soccer-prefix) — IndexedDB keys are
  opaque; renaming requires a migration and adds zero functional value.
- Nit 9 (migration 030 verification ref pointer) — cosmetic.

Pre-existing flake watch

- TeamManagerModal.test.tsx "blocks team unarchive when free limit is
  reached" failed on the previous push, no code in this PR touches
  that file. Pushing this commit re-runs CI; if the flake repeats it
  needs an act() wrap (separate concern, separate PR).

5 files, +51/−4.

+++ src/utils/appStateSchema.ts
@@ -40,7 +40,10 @@ export const gameEventSchema = z.object({
 
 export const scheduledSubSchema = z.object({
   id: z.string().min(1),
-  timeSeconds: z.number().min(0),
+  // Integer constraint matches validateScheduledSubs (validation.ts):
+  // a backup with fractional timeSeconds would pass this schema but
+  // fail the next DataStore save, leaving the restored game unsavable.
+  timeSeconds: z.number().int().min(0),
   outPlayer: z.string().min(1),
   inPlayer: z.string().min(1),
   positionRole: z.string(),
